+++ api/src/pdc/services/acquisition/actions/CancelJourneyAction.ts
@@ -3,8 +3,8 @@ import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { copyGroupIdAndApplyGroupPermissionMiddlewares } from "@/pdc/providers/middleware/index.ts";
 
 import { CarpoolAcquisitionService } from "@/pdc/providers/carpool/index.ts";
-import { handlerConfig, ParamsInterface, ResultInterface } from "../interfaces/cancel.contract.ts";
-import { alias } from "../interfaces/cancel.schema.ts";
+import { handlerConfig, ParamsInterface, ResultInterface } from "../contracts/cancel.contract.ts";
+import { alias } from "../contracts/cancel.schema.ts";
 
 @handler({
   ...handlerConfig,
