+++ api/src/pdc/services/acquisition/actions/PatchJourneyAction.ts
@@ -4,8 +4,8 @@ import { CarpoolAcquisitionService } from "@/pdc/providers/carpool/index.ts";
 import { OperatorClass } from "@/pdc/providers/carpool/interfaces/index.ts";
 import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
-import { handlerConfig, ParamsInterface, ResultInterface } from "../interfaces/patch.contract.ts";
-import { alias } from "../interfaces/patch.schema.ts";
+import { handlerConfig, ParamsInterface, ResultInterface } from "../contracts/patch.contract.ts";
+import { alias } from "../contracts/patch.schema.ts";
 
 import { get } from "@/lib/object/index.ts";
