+++ api/src/pdc/services/acquisition/contracts/common/interfaces/AcquisitionErrorInterface.ts
