+++ api/src/pdc/services/acquisition/contracts/common/interfaces/AcquisitionInterface.ts
