+++ api/src/pdc/services/acquisition/contracts/common/schemas/createJourneyCommon.ts
