+++ api/src/pdc/services/acquisition/contracts/common/schemas/createJourneySchemaV3.ts
